refactor: auth header 설정 및 관리 utils로 분리
