+++ module-api/src/main/java/com/yapp/sharefood/auth/controller/AuthController.java
@@ -4,6 +4,7 @@ import com.yapp.sharefood.auth.dto.OAuthDto;
 import com.yapp.sharefood.auth.dto.request.AuthCreationRequestDto;
 import com.yapp.sharefood.auth.dto.request.AuthRequestDto;
 import com.yapp.sharefood.auth.service.AuthService;
+import com.yapp.sharefood.auth.utils.AuthUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,14 +19,12 @@ import java.net.URI;
 @RestController
 @RequiredArgsConstructor
 public class AuthController {
-    private static final String AUTH_TOKEN_HEADER = "Authorization";
-
     private final AuthService authService;
 
     @PostMapping("/api/v1/auth")
     public ResponseEntity<Void> authenticate(@RequestBody @Valid AuthRequestDto authRequestDto, HttpServletResponse response) {
         OAuthDto oauthDto = authService.authenticate(authRequestDto);
-        response.setHeader(AUTH_TOKEN_HEADER, oauthDto.getToken());
+        AuthUtils.setTokenInHeader(response, oauthDto.getToken());
 
         return ResponseEntity.ok().build();
     }
@@ -33,7 +32,7 @@ public class AuthController {
     @PostMapping("/api/v1/auth/creation")
     public ResponseEntity<URI> signUp(@RequestBody @Valid AuthCreationRequestDto creationRequestDto, HttpServletResponse response) {
         OAuthDto oauthDto = authService.singUp(creationRequestDto);
-        response.setHeader(AUTH_TOKEN_HEADER, oauthDto.getToken());
+        AuthUtils.setTokenInHeader(response, oauthDto.getToken());
 
         URI userCreateUri = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
